Further reduce boost usage
Removes every boost::algorithm, boost::exceptions and boost::assign.
Removes standard.h in favour of using size_t.

+++ cpp/ycm/ClangCompleter/ClangHelpers.cpp
@@ -15,7 +15,6 @@
 // You should have received a copy of the GNU General Public License
 // along with ycmd.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "standard.h"
 #include "ClangHelpers.h"
 #include "ClangUtils.h"
 #include "Utils.h"
@@ -53,7 +52,7 @@ FixIt BuildFixIt( const std::string& text,
                   CXDiagnostic diagnostic ) {
   FixIt fixit;
 
-  uint num_chunks = clang_getDiagnosticNumFixIts( diagnostic );
+  size_t num_chunks = clang_getDiagnosticNumFixIts( diagnostic );
   if ( !num_chunks )
     return fixit;
 
@@ -61,7 +60,7 @@ FixIt BuildFixIt( const std::string& text,
   fixit.location = Location( clang_getDiagnosticLocation( diagnostic ) );
   fixit.text = text;
 
-  for ( uint idx = 0; idx < num_chunks; ++idx ) {
+  for ( size_t idx = 0; idx < num_chunks; ++idx ) {
     FixItChunk chunk;
     CXSourceRange sourceRange;
     chunk.replacement_text = CXStringToString(
@@ -107,12 +106,12 @@ void BuildFullDiagnosticDataFromChildren(
   if ( !diag_set )
     return;
 
-  uint num_child_diagnostics = clang_getNumDiagnosticsInSet( diag_set );
+  size_t num_child_diagnostics = clang_getNumDiagnosticsInSet( diag_set );
 
   if ( !num_child_diagnostics )
     return;
 
-  for ( uint i = 0; i < num_child_diagnostics; ++i ) {
+  for ( size_t i = 0; i < num_child_diagnostics; ++i ) {
     CXDiagnostic child_diag = clang_getDiagnosticInSet( diag_set, i );
 
     if( !child_diag )
@@ -141,10 +140,10 @@ bool CompletionStringAvailable( CXCompletionString completion_string ) {
 
 std::vector< Range > GetRanges( const DiagnosticWrap &diagnostic_wrap ) {
   std::vector< Range > ranges;
-  uint num_ranges = clang_getDiagnosticNumRanges( diagnostic_wrap.get() );
+  size_t num_ranges = clang_getDiagnosticNumRanges( diagnostic_wrap.get() );
   ranges.reserve( num_ranges );
 
-  for ( uint i = 0; i < num_ranges; ++i ) {
+  for ( size_t i = 0; i < num_ranges; ++i ) {
     ranges.push_back(
       Range( clang_getDiagnosticRange( diagnostic_wrap.get(), i ) ) );
   }
@@ -165,13 +164,13 @@ Range GetLocationExtent( CXSourceLocation source_location,
   CXSourceRange range = clang_getCursorExtent(
                           clang_getCursor( translation_unit, source_location ) );
   CXToken *tokens;
-  uint num_tokens;
+  unsigned int num_tokens;
   clang_tokenize( translation_unit, range, &tokens, &num_tokens );
 
   Location location( source_location );
   Range final_range;
 
-  for ( uint i = 0; i < num_tokens; ++i ) {
+  for ( size_t i = 0; i < num_tokens; ++i ) {
     Location token_location( clang_getTokenLocation( translation_unit,
                                                      tokens[ i ] ) );
 
@@ -196,7 +195,7 @@ std::vector< CXUnsavedFile > ToCXUnsavedFiles(
   const std::vector< UnsavedFile > &unsaved_files ) {
   std::vector< CXUnsavedFile > clang_unsaved_files( unsaved_files.size() );
 
-  for ( uint i = 0; i < unsaved_files.size(); ++i ) {
+  for ( size_t i = 0; i < unsaved_files.size(); ++i ) {
     clang_unsaved_files[ i ].Filename = unsaved_files[ i ].filename_.c_str();
     clang_unsaved_files[ i ].Contents = unsaved_files[ i ].contents_.c_str();
     clang_unsaved_files[ i ].Length   = unsaved_files[ i ].length_;
@@ -214,16 +213,16 @@ std::vector< CompletionData > ToCompletionDataVector(
     return completions;
 
   completions.reserve( results->NumResults );
-  unordered_map< std::string, uint > seen_data;
+  unordered_map< std::string, size_t > seen_data;
 
-  for ( uint i = 0; i < results->NumResults; ++i ) {
+  for ( size_t i = 0; i < results->NumResults; ++i ) {
     CXCompletionResult completion_result = results->Results[ i ];
 
     if ( !CompletionStringAvailable( completion_result.CompletionString ) )
       continue;
 
     CompletionData data( completion_result );
-    uint index = GetValueElseInsert( seen_data,
+    size_t index = GetValueElseInsert( seen_data,
                                      data.original_string_,
                                      completions.size() );
